refactor (webapi): Application startup actor
