+++ webapi/src/main/resources/application.conf
@@ -151,7 +151,7 @@ akka {
 
 akka.actor.deployment {
 
-    /storeManager/triplestoreManager/httpTriplestoreRouter {
+    /applicationManager/storeManager/triplestoreManager/httpTriplestoreRouter {
         router = balancing-pool
         nr-of-instances = 10
         pool-dispatcher {
@@ -165,7 +165,7 @@ akka.actor.deployment {
         }
     }
 
-    /storeManager/iiifManager/sipiConnector {
+    /applicationManager/storeManager/iiifManager/sipiConnector {
         router = balancing-pool
         nr-of-instances = 10
         pool-dispatcher {
