+++ webapi/src/main/scala/org/knora/webapi/KnoraService.scala
@@ -20,30 +20,14 @@
 package org.knora.webapi
 
 import akka.actor._
-import akka.http.scaladsl.Http
-import akka.http.scaladsl.server.Directives._
-import akka.http.scaladsl.server.Route
 import akka.stream.ActorMaterializer
-import akka.util.Timeout
-import com.typesafe.scalalogging.Logger
-import kamon.Kamon
 import org.knora.webapi.app._
-import org.knora.webapi.http.CORSSupport.CORS
-import org.knora.webapi.http.ServerVersion.addServerHeader
-import org.knora.webapi.messages.app.appmessages._
-import org.knora.webapi.responders._
-import org.knora.webapi.routing._
-import org.knora.webapi.routing.admin._
-import org.knora.webapi.routing.v1._
-import org.knora.webapi.routing.v2._
-import org.knora.webapi.store._
-import org.knora.webapi.util.{CacheUtil, StringFormatter}
+import org.knora.webapi.util.StringFormatter
 
-import scala.concurrent.duration._
-import scala.concurrent.{Await, ExecutionContext, Future}
+import scala.concurrent.ExecutionContext
 import scala.language.postfixOps
 import scala.languageFeature.postfixOps
-import scala.util.{Failure, Success}
+
 
 /**
   * Knora Core abstraction.
@@ -85,136 +69,16 @@ trait LiveCore extends Core {
 }
 
 /**
-  * Provides methods for starting and stopping Knora from within another application. This is where the actor system
-  * along with the three main supervisor actors is started. All further actors are started and supervised by those
-  * three actors.
+  * Starts the main application actor
   */
-trait KnoraService extends AroundDirectives {
+trait KnoraService {
     this: Core =>
 
     // Initialise StringFormatter with the system settings. This must happen before any responders are constructed.
     StringFormatter.init(settings)
 
-    import scala.language.postfixOps
-
     /**
       * The actor used at startup, transitioning between states, and storing the application application wide variables in a thread safe manner.
       */
-    protected val applicationStateActor: ActorRef = system.actorOf(Props(new ApplicationStateActor).withDispatcher(KnoraDispatchers.KnoraActorDispatcher), name = APPLICATION_STATE_ACTOR_NAME)
-
-
-    // #supervisors
-    /**
-      * The supervisor actor that forwards messages to actors that deal with persistent storage.
-      */
-    protected val storeManager: ActorRef = system.actorOf(Props(new StoreManager with LiveActorMaker).withDispatcher(KnoraDispatchers.KnoraActorDispatcher), name = StoreManagerActorName)
-
-
-    /**
-      * The supervisor actor that forwards messages to responder actors to handle API requests.
-      */
-    protected val responderManager: ActorRef = system.actorOf(Props(new ResponderManager(applicationStateActor, storeManager) with LiveActorMaker).withDispatcher(KnoraDispatchers.KnoraActorDispatcher), name = RESPONDER_MANAGER_ACTOR_NAME)
-    // #supervisors
-
-    /**
-      * Timeout definition
-      */
-    implicit protected val timeout: Timeout = settings.defaultTimeout
-
-    /**
-      * A user representing the Knora API server, used for initialisation on startup.
-      */
-    private val systemUser = KnoraSystemInstances.Users.SystemUser
-
-    /**
-      * Route data.
-      */
-    private val routeData = KnoraRouteData(system, applicationStateActor, responderManager, storeManager)
-
-    private val logger = Logger("KnoraService")
-
-    /**
-      * All routes composed together and CORS activated.
-      */
-    private val apiRoutes: Route = logDuration(logger) {
-        addServerHeader {
-            CORS(
-                  new HealthRoute(routeData).knoraApiPath ~
-                  new RejectingRoute(routeData).knoraApiPath ~
-                  new ClientApiRoute(routeData).knoraApiPath ~
-                  new ResourcesRouteV1(routeData).knoraApiPath ~
-                  new ValuesRouteV1(routeData).knoraApiPath ~
-                  new StandoffRouteV1(routeData).knoraApiPath ~
-                  new ListsRouteV1(routeData).knoraApiPath ~
-                  new ResourceTypesRouteV1(routeData).knoraApiPath ~
-                  new SearchRouteV1(routeData).knoraApiPath ~
-                  new AuthenticationRouteV1(routeData).knoraApiPath ~
-                  new AssetsRouteV1(routeData).knoraApiPath ~
-                  new CkanRouteV1(routeData).knoraApiPath ~
-                  new UsersRouteV1(routeData).knoraApiPath ~
-                  new ProjectsRouteV1(routeData).knoraApiPath ~
-                  new OntologiesRouteV2(routeData).knoraApiPath ~
-                  new SearchRouteV2(routeData).knoraApiPath ~
-                  new ResourcesRouteV2(routeData).knoraApiPath ~
-                  new ValuesRouteV2(routeData).knoraApiPath ~
-                  new StandoffRouteV2(routeData).knoraApiPath ~
-                  new ListsRouteV2(routeData).knoraApiPath ~
-                  new AuthenticationRouteV2(routeData).knoraApiPath ~
-                  new GroupsRouteADM(routeData).knoraApiPath ~
-                  new ListsRouteADM(routeData).knoraApiPath ~
-                  new PermissionsRouteADM(routeData).knoraApiPath ~
-                  new ProjectsRouteADM(routeData).knoraApiPath ~
-                  new StoreRouteADM(routeData).knoraApiPath ~
-                  new UsersRouteADM(routeData).knoraApiPath ~
-                  new SipiRouteADM(routeData).knoraApiPath ~
-                  new SwaggerApiDocsRoute(routeData).knoraApiPath,
-                settings
-            )
-        }
-    }
-
-    // #startService
-    /**
-      * Starts the Knora API server.
-      */
-    def startService(skipLoadingOfOntologies: Boolean): Unit = {
-
-        val bindingFuture: Future[Http.ServerBinding] = Http().bindAndHandle(Route.handlerFlow(apiRoutes), settings.internalKnoraApiHost, settings.internalKnoraApiPort)
-
-        bindingFuture onComplete {
-            case Success(_) => {
-
-                if (settings.prometheusEndpoint) {
-                    // Load Kamon monitoring
-                    Kamon.loadModules()
-                }
-
-                // Kick of startup procedure.
-                applicationStateActor ! InitStartUp(skipLoadingOfOntologies)
-            }
-            case Failure(ex) => {
-                logger.error("Failed to bind to {}:{}! - {}", settings.internalKnoraApiHost, settings.internalKnoraApiPort, ex.getMessage)
-                stopService()
-            }
-        }
-    }
-    // #startService
-
-    /**
-      * Stops Knora.
-      */
-    def stopService(): Unit = {
-        logger.info("KnoraService - Shutting down.")
-
-        if (settings.prometheusEndpoint) {
-            // Stop Kamon monitoring
-            Kamon.stopModules()
-        }
-
-        Http().shutdownAllConnectionPools()
-        CacheUtil.removeAllCaches()
-
-        system.terminate()
-        Await.result(system.whenTerminated, 30 seconds)
-    }
+    protected val appActor: ActorRef = system.actorOf(Props(new ApplicationActor with LiveManagers).withDispatcher(KnoraDispatchers.KnoraActorDispatcher), name = APPLICATION_MANAGER_ACTOR_NAME)
 }
