+++ webapi/src/main/scala/org/knora/webapi/Main.scala
@@ -19,7 +19,11 @@
 
 package org.knora.webapi
 
+import akka.actor.Terminated
 import org.knora.webapi.messages.app.appmessages._
+
+import scala.concurrent.duration._
+import scala.concurrent.{Await, Future}
 
 /**
   * Starts Knora by bringing everything into scope by using the cake pattern. The [[LiveCore]] trait provides
@@ -31,14 +35,14 @@ object Main extends App with LiveCore with KnoraService {
     val arglist = args.toList
 
     // loads demo data
-    if (arglist.contains("loadDemoData")) applicationStateActor ! SetLoadDemoDataState(true)
-    if (arglist.contains("--load-demo-data")) applicationStateActor ! SetLoadDemoDataState(true)
-    if (arglist.contains("-d")) applicationStateActor ! SetLoadDemoDataState(true)
+    if (arglist.contains("loadDemoData")) appActor ! SetLoadDemoDataState(true)
+    if (arglist.contains("--load-demo-data")) appActor ! SetLoadDemoDataState(true)
+    if (arglist.contains("-d")) appActor ! SetLoadDemoDataState(true)
 
     // allows reloading of data over HTTP
-    if (arglist.contains("allowReloadOverHTTP")) applicationStateActor ! SetAllowReloadOverHTTPState(true)
-    if (arglist.contains("--allow-reload-over-http")) applicationStateActor ! SetAllowReloadOverHTTPState(true)
-    if (arglist.contains("-r")) applicationStateActor ! SetAllowReloadOverHTTPState(true)
+    if (arglist.contains("allowReloadOverHTTP")) appActor ! SetAllowReloadOverHTTPState(true)
+    if (arglist.contains("--allow-reload-over-http")) appActor ! SetAllowReloadOverHTTPState(true)
+    if (arglist.contains("-r")) appActor ! SetAllowReloadOverHTTPState(true)
 
     if (arglist.contains("--help")) {
         println(
@@ -59,9 +63,23 @@ object Main extends App with LiveCore with KnoraService {
     } else {
         /* Start the HTTP layer, allowing access */
         /* Don't skip loading of ontologies */
-        startService(skipLoadingOfOntologies=false)
+        appActor ! AppStart(skipLoadingOfOntologies=false)
 
-        /* add the method for shutting down our application to the shutdown hook, so that we can clean up */
-        scala.sys.addShutdownHook(stopService())
+        /**
+          * Adds shutting down of our actor system to the shutdown hook.
+          * Because we are blocking, we will run this on a separate thread.
+          */
+        scala.sys.addShutdownHook(
+            new Thread(
+                () => {
+                    val terminate: Future[Terminated] = system.terminate()
+                    Await.result(terminate, 30.seconds)
+                }
+            )
+        )
+
+        system.registerOnTermination {
+            println("ActorSystem terminated")
+        }
     }
 }
