+++ webapi/src/main/scala/org/knora/webapi/app/package.scala
@@ -2,7 +2,7 @@ package org.knora.webapi
 
 package object app {
 
-    val APPLICATION_STATE_ACTOR_NAME = "applicationState"
-    val APPLICATION_STATE_ACTOR_PATH = "/user/" + APPLICATION_STATE_ACTOR_NAME
+    val APPLICATION_MANAGER_ACTOR_NAME = "applicationManager"
+    val APPLICATION_MANAGER_ACTOR_PATH = "/user/" + APPLICATION_MANAGER_ACTOR_NAME
 
 }
