+++ webapi/src/main/scala/org/knora/webapi/responders/v2/search/gravsearch/types/InferringGravsearchTypeInspector.scala
@@ -41,7 +41,7 @@ class InferringGravsearchTypeInspector(nextInspector: Option[GravsearchTypeInspe
 
     import InferringGravsearchTypeInspector._
 
-    private val responderManager = responderData.responderManager
+    private val responderManager = responderData.appActor
 
     private implicit val stringFormatter: StringFormatter = StringFormatter.getGeneralInstance
