+++ webapi/src/test/resources/logback-test.xml
@@ -54,7 +54,7 @@
     <logger name="org.knora" level="INFO"/>
     <logger name="org.knora.webapi" level="INFO"/>
     <logger name="org.knora.webapi.KnoraExceptionHandler" level="INFO"/>
-    <logger name="org.knora.webapi.app.ApplicationStateActor" level="INFO"/>
+    <logger name="org.knora.webapi.app.ApplicationActor" level="INFO"/>
     <logger name="org.knora.webapi.util.cache" level="INFO"/>
     <logger name="org.knora.webapi.util.PermissionUtilADM" level="INFO"/>
     <logger name="org.knora.webapi.util.TransformData" level="INFO"/>
