+++ webapi/src/test/scala/org/knora/webapi/E2ESimSpec.scala
@@ -28,7 +28,7 @@ import akka.stream.ActorMaterializer
 import com.typesafe.config.{Config, ConfigFactory}
 import com.typesafe.scalalogging.LazyLogging
 import io.gatling.core.scenario.Simulation
-import org.knora.webapi.messages.app.appmessages.SetAllowReloadOverHTTPState
+import org.knora.webapi.messages.app.appmessages.{AppStart, AppStop, SetAllowReloadOverHTTPState}
 import org.knora.webapi.messages.store.triplestoremessages.{RdfDataObject, TriplestoreJsonProtocol}
 import org.knora.webapi.util.StringFormatter
 import spray.json._
@@ -55,26 +55,20 @@ object E2ESimSpec {
   */
 abstract class E2ESimSpec(_system: ActorSystem) extends Simulation with Core with KnoraService with TriplestoreJsonProtocol with RequestBuilding with LazyLogging {
 
-    /* needed by the core trait */
-
-    implicit lazy val settings: SettingsImpl = Settings(system)
-
-    implicit val materializer: ActorMaterializer = ActorMaterializer()
-
-    implicit val executionContext: ExecutionContext = system.dispatchers.defaultGlobalDispatcher
-
-    StringFormatter.initForTest()
-
+    /* constructors */
     def this(name: String, config: Config) = this(ActorSystem(name, config.withFallback(E2ESimSpec.defaultConfig)))
-
     def this(config: Config) = this(ActorSystem("PerfSpec", config.withFallback(E2ESimSpec.defaultConfig)))
-
     def this(name: String) = this(ActorSystem(name, E2ESimSpec.defaultConfig))
-
     def this() = this(ActorSystem("PerfSpec", E2ESimSpec.defaultConfig))
 
     /* needed by the core trait */
     implicit lazy val system: ActorSystem = _system
+    implicit lazy val settings: SettingsImpl = Settings(system)
+    implicit val materializer: ActorMaterializer = ActorMaterializer()
+    implicit val executionContext: ExecutionContext = system.dispatchers.defaultGlobalDispatcher
+
+    /* Needs to be initialized before any responders */
+    StringFormatter.initForTest()
 
     protected val baseApiUrl: String = settings.internalKnoraApiBaseUrl
 
@@ -85,12 +79,10 @@ abstract class E2ESimSpec(_system: ActorSystem) extends Simulation with Core wit
         /* Set the startup flags and start the Knora Server */
         logger.info(s"executing before setup started")
 
-        applicationStateActor ! SetAllowReloadOverHTTPState(true)
+        appActor ! SetAllowReloadOverHTTPState(true)
 
-        // start the knora service without startupTasks
-        startService(false)
+        appActor ! AppStart(false)
 
-        // loadTestData
         loadTestData(rdfDataObjects)
 
         logger.info(s"executing before setup finished")
@@ -99,7 +91,7 @@ abstract class E2ESimSpec(_system: ActorSystem) extends Simulation with Core wit
     after {
         /* Stop the server when everything else has finished */
         logger.info(s"executing after setup")
-        stopService()
+        appActor ! AppStop()
     }
 
     protected def loadTestData(rdfDataObjects: Seq[RdfDataObject]): Unit = {
